+++ Sources/Extensions/NSUserDefaults+Subscripts.swift
@@ -1,5 +1,5 @@
 //
-// SwiftyUserDefaults
+// Transient
 //
 // Copyright (c) 2015-2016 Radosław Pietruszewski
 //